+++ dev/snippets/button/group_justified.cljs
@@ -1,11 +1,10 @@
 #_
 (:require [om-bootstrap.button :as b])
 
-(b/toolbar {}
-           (b/button {} "Default")
-           (b/button {:bs-style "primary"} "Primary")
-           (b/button {:bs-style "success"} "Success")
-           (b/button {:bs-style "info"} "Info")
-           (b/button {:bs-style "warning"} "Warning")
-           (b/button {:bs-style "danger"} "Danger")
-           (b/button {:bs-style "link"} "Link"))
+(b/button-group
+ {:justified? true}
+ (b/button {:href "#"} "Left")
+ (b/button {:href "#"} "Middle")
+ (b/dropdown {:title "Dropdown"}
+             (b/menu-item {:key 1} "Dropdown link")
+             (b/menu-item {:key 2} "Dropdown link")))